+++ lib/src/request_details.dart
@@ -1,3 +1,6 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+import 'dart:convert';
+
 import 'package:requests_inspector/requests_inspector.dart';
 
 class RequestDetails {
@@ -88,4 +91,38 @@ class RequestDetails {
         responseBody.hashCode ^
         sentTime.hashCode;
   }
+
+  Map<String, dynamic> toMap() {
+    return <String, dynamic>{
+      'requestName': requestName,
+      'requestMethod': requestMethod.name,
+      'url': url,
+      'statusCode': statusCode,
+      'headers': headers,
+      'queryParameters': queryParameters,
+      'requestBody': requestBody,
+      'responseBody': responseBody,
+      'sentTime': sentTime.toIso8601String(),
+    }..removeWhere((_, v) => v == null);
+  }
+
+  factory RequestDetails.fromMap(Map<String, dynamic> map) {
+    return RequestDetails(
+      requestName: map['requestName'],
+      requestMethod: RequestMethod.values
+          .firstWhere((e) => e.name == map['requestMethod']),
+      url: map['url'] as String,
+      statusCode: map['statusCode'] != null ? map['statusCode'] as int : null,
+      headers: map['headers'] as dynamic,
+      queryParameters: map['queryParameters'] as dynamic,
+      requestBody: map['requestBody'] as dynamic,
+      responseBody: map['responseBody'] as dynamic,
+      sentTime: DateTime.parse(map['sentTime']),
+    );
+  }
+
+  String toJson() => json.encode(toMap());
+
+  factory RequestDetails.fromJson(String source) =>
+      RequestDetails.fromMap(json.decode(source) as Map<String, dynamic>);
 }